% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManuipulateDataset.R
\name{AllMetsPresent}
\alias{AllMetsPresent}
\title{Filters a data frame in wide format that contains metabolites present
in every sample}
\usage{
AllMetsPresent(cMetDataWide)
}
\arguments{
\item{cMetDataWide}{a count matrix of Metabolon data in Wide format}
}
\value{
a Data Frame which has metabolites missing in *any* sample removed
}
\description{
Filters a data frame in wide format that contains metabolites present
in every sample
}
\examples{
allMetsPresentWide <- AllMetsPresent(metDataWide)
}

