% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandyUtilities.R
\name{AsNumericFactor}
\alias{AsNumericFactor}
\title{Converts a factor with numeric character levels into a factor with numeric levels}
\usage{
AsNumericFactor(x)
}
\arguments{
\item{x}{- a factor with numeric character levels}
}
\value{
a factor with levels that are numeric, not character
}
\description{
Converts a factor with numeric character levels into a factor with numeric levels
}
\examples{
TBD
}

