% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormaliseDataset.R
\name{AssignLabelsToGroups}
\alias{AssignLabelsToGroups}
\title{Convenience function for the median normalisation.}
\usage{
AssignLabelsToGroups(groupList, valInGroup)
}
\arguments{
\item{groupList}{A list of arrays of groups}

\item{valInGroup}{An value to find in one of the arrays in the groupList}
}
\value{
the index in the list that valInGroup belongs to
}
\description{
The function assigns a specific value to members of a specific group. Used in order
to group Metabolon run days together, e.g. 1, 2, 3 are actually the same run day.
}
\examples{
AssignLabelsToGroups(list(c(1, 2, 3), c(4, 5, 6)), 6)
#returns 2 as item 6 is in the second group in the list
}

