% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManuipulateDataset.R
\name{CalculateMetPCA}
\alias{CalculateMetPCA}
\title{Calculates a scaled and centred PCA analysis of a Metabolon dataset}
\usage{
CalculateMetPCA(cMetDataLong)
}
\arguments{
\item{cMetDataLong}{a count dataframe of Metabolon data in Long format}

\item{sampleInfo}{a dataframe of sample metadata}
}
\value{
an S3 pca object
}
\description{
Calculates a scaled and centred PCA analysis of a Metabolon dataset
}
\examples{
wideMetCounts <- CalculateMetPCA(longMetCounts)
}

