% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormaliseDataset.R
\name{CCMNNormaliseDataset}
\alias{CCMNNormaliseDataset}
\title{Normalises a Metabolon dataset using the Cross-contribution Compensating
Multiple Internal standard normalistaion" (CCMN) method, Redestig et al. 2009}
\usage{
CCMNNormaliseDataset(metInfo, sampleInfo, metDataLong, biologicalFactor,
  matchOrder = NULL)
}
\arguments{
\item{metInfo}{A dataframe of metabolite metadata (must include Internal Standards)}

\item{sampleInfo}{A dataframe containing the sample metadata}

\item{metDataLong}{A dataframe of metabolite data in Long Format}

\item{biologicalFactor}{A chararacter factor in the samples metadata to calulcate cross-contamination with}

\item{matchOrder}{flag for re-ordering}
}
\description{
Normalises a Metabolon dataset using the Cross-contribution Compensating
Multiple Internal standard normalistaion" (CCMN) method, Redestig et al. 2009
}

