% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandyUtilities.R
\name{CoVar}
\alias{CoVar}
\title{Calculates the percentage Coefficent of Variation from a numeric vector}
\usage{
CoVar(x)
}
\arguments{
\item{x}{- numeric vector}
}
\value{
% Coefficient of Variation of the vector x
}
\description{
Calculates the percentage Coefficent of Variation from a numeric vector
}
\examples{
CoVar(c(2, 1, 3)) returns 50
}

