% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandyUtilities.R
\name{CreateOrClean}
\alias{CreateOrClean}
\title{creates or cleans a given directory (non recursively). If the dir doesn't exist
the function will create it, if it does exist, the function will remove all items
from the directory. Used to create ''Output'' directories whose items are expendible}
\usage{
CreateOrClean(directory)
}
\arguments{
\item{directory}{- the dir to create or to remove all items from non recursively}
}
\description{
creates or cleans a given directory (non recursively). If the dir doesn't exist
the function will create it, if it does exist, the function will remove all items
from the directory. Used to create ''Output'' directories whose items are expendible
}
\examples{
CreateOrClean('path/to/my/directory')

}

