% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManuipulateDataset.R
\name{FilterMissingMets}
\alias{FilterMissingMets}
\title{Filters a metabolite count data frame in long format.  Removes metabolites missing in samples at a greater
frequency than a threshold set by the user}
\usage{
FilterMissingMets(cMetDataLong, proportion)
}
\arguments{
\item{cMetDataLong}{a count matrix of Metabolon data in Long format}

\item{cMetDataLong}{proportion of metabolite frequency to filter out}
}
\value{
a Data Frame which has metabolites present at a lower frequency than 'proportion' removed
}
\description{
Filters a metabolite count data frame in long format.  Removes metabolites missing in samples at a greater
frequency than a threshold set by the user
}
\examples{
allMetsPresentLong <- FilterMissingMets(cMetDataLong, 0.2)
removes all metabolites missing in more than 20\% of the samples
}

