% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandyUtilities.R
\name{HasNA}
\alias{HasNA}
\title{Calculates the number of NA entries in a vector x}
\usage{
HasNA(x)
}
\arguments{
\item{x}{- a vector of numeric values, potentially with some NAs}
}
\value{
sum of NA entries in x
}
\description{
Calculates the number of NA entries in a vector x
}
\examples{
HasNA(c(3,4,NA,8))
}

