% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDataset.R
\name{LoadMetabolonData}
\alias{LoadMetabolonData}
\title{Load Metabolon datasets from preconstructed sample file, metabolite info file and a data file}
\usage{
LoadMetabolonData(sampleFile, metaboliteFile, dataFile)
}
\arguments{
\item{sampleFile}{A filename containing information about the samples}

\item{metaboliteFile}{A filename containing information about the detected metabolites}

\item{dataFile}{A filename containing the data, typically raw ion-counts, for the samples and metabolites}
}
\value{
A list of DataFrames containing the sample, metabolite and count information
}
\description{
Load Metabolon datasets from preconstructed sample file, metabolite info file and a data file
}
\examples{
metabolonDataSet <- ReadMetabolonData('sampleInfo.csv', 'metaboliteInfo.csv', 'data.csv')
}

