% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandyUtilities.R
\name{MeanCentred}
\alias{MeanCentred}
\title{Removes the mean from a numeric vector, removes NAs}
\usage{
MeanCentred(x)
}
\arguments{
\item{x}{- numeric vector}
}
\value{
a mean centred numeric vector
}
\description{
Removes the mean from a numeric vector, removes NAs
}
\examples{
MeanCentered(c(2, 1, 3)) returns c(0, -1, 1)

}

