% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormaliseDataset.R
\name{MedianNormaliseDataset}
\alias{MedianNormaliseDataset}
\title{Performs Metabolon median normalisation, typically by 'Run Day'}
\usage{
MedianNormaliseDataset(cSampleInfo, cMetLog10DataLong, normaliseField,
  normaliseGroups, scaleToMedianRawCount = TRUE)
}
\arguments{
\item{cSampleInfo}{A DataFrame of sample information}

\item{cMetLog10DataLong}{A DataFrame of log 10 Metabolite Counts in Long Format}

\item{normaliseField}{A field in cSampleInfo over which the Normalisation should be performed}

\item{normaliseGroups}{A list of groupings of the levels of normaliseField to treat as normalisation blocks}

\item{scaleToMedianRawCount}{A boolean that asks whether to adjust to the global median metabolite level
or to adjust to a median of 1}

\item{cMetInfo}{A DataFrame of metabolite information}
}
\value{
A dataframe in Long format normalised by ''normaliseField''.
}
\description{
Performs Metabolon median normalisation, typically by 'Run Day'
}

