% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSets.R
\docType{data}
\name{metaboliteDataWide}
\alias{metaboliteDataWide}
\title{Test Sample Infomation}
\format{a \code{DataFrame} instance, 1 row per sample that gives counts for each of the metabolites}
\source{
Randomly sampled Metabolon data
}
\usage{
metaboliteDataWide
}
\description{
Contains metabolite counts of the 29 test metabolites on the 20 samples
}
\author{
Michael Epstein
}
\keyword{datasets}

