% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSets.R
\docType{data}
\name{metaboliteInfo}
\alias{metaboliteInfo}
\title{Test Metabolite Infomation}
\format{a \code{DataFrame} instance, 1 row per metabolite detailing its chemical information}
\source{
Fields taken from \code{Metabolon Inc.}
}
\usage{
metaboliteInfo
}
\description{
Contains metadata on 29 test metabolites with 15 fields
}
\author{
Michael Epstein
}
\keyword{datasets}

