% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManuipulateDataset.R
\name{MetDataLongToWide}
\alias{MetDataLongToWide}
\title{Convenience function for transforming a Long count Data Frame to a Wide count Data Frame}
\usage{
MetDataLongToWide(cMetDataLong)
}
\arguments{
\item{cMetDataLong}{a count matrix of Metabolon data in Long format}
}
\value{
a count matrix of Metabolon data in Wide format
}
\description{
Convenience function for transforming a Long count Data Frame to a Wide count Data Frame
}
\examples{
wideMetCounts <- MetDataLongToWide(longMetCounts)
}

