% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormaliseDataset.R
\name{NOMISNormaliseDataset}
\alias{NOMISNormaliseDataset}
\title{Normalises a Metabolon dataset using the "Normalization using Optimal
selection of Multiple Internal Standards" method of Sysi-Aho et al 2007}
\usage{
NOMISNormaliseDataset(metInfo, metDataLong)
}
\arguments{
\item{metInfo}{A dataframe of metabolite metadata (must include the Internal Standards)}

\item{metDataLong}{A dataframe of metabolite data in Long Format}
}
\value{
A dataframe of NOMIS normalised data in Long Format
}
\description{
Normalises a Metabolon dataset using the "Normalization using Optimal
selection of Multiple Internal Standards" method of Sysi-Aho et al 2007
}

