% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandyUtilities.R
\name{RenameColumns}
\alias{RenameColumns}
\title{Renames column names in a dataframe, so each column name matching
columnsBy[[i]] becomes columnsTo[[i]]}
\usage{
RenameColumns(df, columnsBy, columnsTo, fill = NA)
}
\arguments{
\item{df}{A data frame with which to rename the colnames}

\item{columnsBy}{A vector of colnames to be replaced}

\item{columnsTo}{A vector of names to replace with}

\item{fill}{Values to insert into the now missing columns}
}
\description{
Renames column names in a dataframe, so each column name matching
columnsBy[[i]] becomes columnsTo[[i]]
}

