% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSets.R
\docType{data}
\name{sampleInfo}
\alias{sampleInfo}
\title{Test Sample Infomation}
\format{a \code{DataFrame} instance, 1 row per sample detailing its sample metadata}
\source{
Fields taken from \code{Metabolon Inc.}
}
\usage{
sampleInfo
}
\description{
Contains metadata on 20 samples; 9 from two individuals and 2 pooled samples
}
\author{
Michael Epstein
}
\keyword{datasets}

