% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManuipulateDataset.R
\name{TransformExp10Mets}
\alias{TransformExp10Mets}
\title{Transforms the metabolite count in long format by exponentiating each value to base 10}
\usage{
TransformExp10Mets(cMetDataLong)
}
\arguments{
\item{cMetDataLong}{a log10 count matrix of Metabolon data in Long format}
}
\value{
a Data Frame where the metabolite counts are original values
}
\description{
Transforms the metabolite count in long format by exponentiating each value to base 10
}
\examples{
expMetCounts <- TransformExp10Mets(logMetCounts)
}

