% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManuipulateDataset.R
\name{TransformLog10Mets}
\alias{TransformLog10Mets}
\title{Transforms the metabolite count in long format by taking log10 of each value}
\usage{
TransformLog10Mets(cMetDataLong)
}
\arguments{
\item{cMetDataLong}{a count matrix of Metabolon data in Long format}
}
\value{
a Data Frame where the metabolite counts are now log10 values
}
\description{
Transforms the metabolite count in long format by taking log10 of each value
}
\examples{
logMetCounts <- TransformLog10Mets(cMetDataLong)
}

