% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManuipulateDataset.R
\name{TransformZscoreMets}
\alias{TransformZscoreMets}
\title{Calculates the z-score of each metabolite count within each metabolite
Typically requires a log-transformed count matrix for normality}
\usage{
TransformZscoreMets(cMetDataLong)
}
\arguments{
\item{cMetDataLong}{a count matrix of Metabolon data in Long format}
}
\value{
a Data Frame where the metabolite counts within each metabolite are z-scores
}
\description{
Calculates the z-score of each metabolite count within each metabolite
Typically requires a log-transformed count matrix for normality
}
\examples{
zMetCounts <- TransformZscoreMets(logLongMetCounts)
}

