% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormaliseDataset.R
\name{VolumeNormaliseDataset}
\alias{VolumeNormaliseDataset}
\title{Adjust the metabolite counts for differing volumes in supplied biological material}
\usage{
VolumeNormaliseDataset(cSampleInfo, cMetInfo, cMetDataLong, normalVolume,
  sampleType)
}
\arguments{
\item{cSampleInfo}{A Data Frame containing cleaned information about the samples}

\item{cMetInfo}{A Data Frame  containing cleaned information about the detected metabolites}

\item{cMetDataLong}{A Data Frame  containing cleaned data raw ion-counts in Long format}

\item{normalVolume}{The volume (in microL) to adjust the ion-counts to}

\item{sampleType}{The labelling of the sampletype to apply the adjustment to - i.e. not MTRX}
}
\value{
A Data Frame containing volume adjusted count information
}
\description{
Adjust the metabolite counts for differing volumes in supplied biological material
}
\examples{
volumeAdjustedData <- VolumeAdjustMetabolonData(sampleInfo, metInfo, metData, normalVolume = 0.95)

}

